requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-api-core==1.17.0
google-api-python-client==1.8.1
google-auth==1.14.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.9
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySide2==5.14.2.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
shiboken2==5.14.2.1
simpleaudio==1.0.4
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.9
